+++ src/main/java/gtclassic/common/tile/GTTileBedrockMiner.java
@@ -212,9 +212,9 @@ public class GTTileBedrockMiner extends TileEntityElecMachine
 					int count = this.getStackInSlot(i).getCount();
 					this.setStackInSlot(i, GTHelperStack.copyWithSize(this.output, count + this.output.getCount()));
 					world.playSound((EntityPlayer) null, pos, SoundEvents.BLOCK_STONE_BREAK, SoundCategory.BLOCKS, 0.6F, 1.0F);
-					tryDamagePipe();
 					GTUtility.exportFromMachineToSide(this, EnumFacing.UP, getOutputSlots());
 				}
+				tryDamagePipe();
 				this.useEnergy(EU_COST);
 				this.getNetwork().updateTileGuiField(this, "energy");
 				this.setActive(true);